// core/user-dashboard/store/comment/action-types.ts
export enum ActionTypes {
  GET_COMMENTS = "GET_COMMENTS",
  GET_COMMENT = "GET_COMMENT",
  CREATE_COMMENT = "CREATE_COMMENT",
  UPDATE_COMMENT = "UPDATE_COMMENT",
  DELETE_COMMENT = "DELETE_COMMENT",
  HARD_DELETE_COMMENT = "HARD_DELETE_COMMENT",
  GET_COMMENTS_BY_POST = "GET_COMMENTS_BY_POST",
  LIKE_COMMENT = "LIKE_COMMENT",
  REPLY_COMMENT = "REPLY_COMMENT",
  DISLIKE_COMMENT = "DISLIKE_COMMENT",
}